Move InventoryGUICurrentItemShowcase into src/, and include/ dirs
