+++ inventory.hpp
@@ -4,54 +4,8 @@
 
 #include <inventory_gui/item_button.hpp>
 
+#include <inventory_gui/current_item_showcase.hpp>
 
-
-class InventoryGUICurrentItemShowcase : public UIWidget
-{
-public:
-  InventoryItem item;
-  UIImage *featured_image;
-  UIText *title_text;
-  UITextBox *description_textbox;
-
-  InventoryGUICurrentItemShowcase(UIWidget *parent)
-    : UIWidget(parent, "InventoryGUICurrentItemShowcase", new UISurfaceAreaBox(SCREEN_W-500, SCREEN_H/2, 600, 400))
-    , item(InventoryItem::Type::EMPTY)
-    , title_text(NULL)
-    , description_textbox(NULL)
-    , featured_image(NULL)
-  {
-    featured_image = new UIImage(this, place.size.x/2, place.size.y/2, item.get_image());
-    title_text = new UIText(this, 20, 20, item.get_name());
-    title_text->place.align.y = 0;
-    title_text->set_font(Framework::font("space age.otf 34"));
-    description_textbox = new UITextBox(this, place.size.x/2, place.size.y*0.75, place.size.x*0.8, place.size.y*0.25, item.get_description()) ;
-    description_textbox->set_text_color(color::black);
-    description_textbox->place.align.x = 0.5;
-    description_textbox->place.align.y = 0;
-  }
-  void on_draw() override
-  {
-    Style::draw_button(Style::NORMAL, place, item.is_empty() ? "- EMPTY -" : "");
-  }
-  void set_item(InventoryItem item)
-  {
-    this->item = item;
-    featured_image->set_bitmap(item.get_image());
-    title_text->set_text(item.get_name());
-    description_textbox->set_text(item.get_description());
-  }
-  void show(float speed=0.5)
-  {
-    if (Logging::at_least(L_VERBOSE)) std::cout << "InventoryGUICurrentItemShowcase.show()" << std::endl;
-    Framework::motion().cmove_to(&place.position.x, SCREEN_W-500, speed);
-  }
-  void hide(float speed=0.5)
-  {
-    if (Logging::at_least(L_VERBOSE)) std::cout << "InventoryGUICurrentItemShowcase.hide()" << std::endl;
-    Framework::motion().cmove_to(&place.position.x, SCREEN_W+500, speed);
-  }
-};
 
 
 class InventoryGUINotification : public UIWidget
